Haladunk

+++ mzProg2021/Balkezes.cs
@@ -9,22 +9,24 @@ namespace mzProg2021
     class Balkezes
     {
         public string nev { get; set; }
-        public int elso { get; set; }
-        public int utolso { get; set; }
-        public int suly { get; set; }
-        public int magassag { get; set; }
+        public double elso { get; set; }
+        public double utolso { get; set; }
+        public double suly { get; set; }
+        public double magassag { get; set; }
 
         public Balkezes(string sor)
         {
             string[] s = sor.Split(';');
             nev = s[0];
-            elso = int.Parse(s[1]);
-            utolso = int.Parse(s[2]);
-            suly = int.Parse(s[3]);
-            magassag = int.Parse(s[4]);
+            elso = double.Parse(s[1]);
+            utolso = double.Parse(s[2]);
+            suly = double.Parse(s[3]);
+            magassag = double.Parse(s[4]);
+
         }
         public Balkezes()
         {
+          
         }
     }
